Refactor typecheck schemas
This commit removes user schema definition and propType function from the User component and places them in the types.jsx file.
